Fix: daemon component error messages from unresolved vars
The file control is processed by all components, but the variable it
referenced was only in in agent bundle. It would be useful if there were
classes that indicated which component was processing the policy. That
could allow choosing inputs based on the component, and potentially
allow a reduction in number of common bundles used simply to allow for
dynamic paths. This is also because the paths are not consistently
relative to the current file. I have seen when running policy with a
path of ./ that can mess up other files relative file control
inputs (needs further qualification). Common bundles can be good for
feature class based dynamic inputs.

Changelog: Title